[fix] extend number of constants per chunk from UINT8_MAX to UINT16_MAX

+++ lib/compiler.c
@@ -141,14 +141,14 @@ static int emitJump(uint8_t instruction) {
 }
 static void emitReturn() { emitByte(OP_RETURN); }
 
-static uint8_t makeConstant(Value value) {
+static uint16_t makeConstant(Value value) {
   int constant = addConstant(currentChunk(), value);
-  if (constant > UINT8_MAX) {
+  if (constant > UINT16_MAX) {
     error("Too many constants in one chunk.");
     return 0;
   }
 
-  return (uint8_t)constant;
+  return (uint16_t)constant;
 }
 
 static void emitConstant(Value value) {
